Add maps card functionality to HomeActivity, UI changes
- Added Google Maps API and meta to manifest
- Got rid of ResideMenu as it gave layout bugs
- Added higher resolution icons
- Twitter card now has date

- Added UI logic for old tweets, before it was always saying
"Here is where we are today", but that wouldn't be accurate
if the tweet wasn't from today. This also made the weatherbar message
very weird. Changed it to figure out if a tweet was posted the same day as the user, so the app displays either
"Here is where we are today" or "Here's where we were last".

- Twiter card have ripple effects when you touch them
- The navbar menu button animates
- Landscape is finally not broken!